+++ app/src/main/java/com/example/sendbacksendbag/Home.kt
@@ -16,7 +16,8 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.navigation.NavController
-import com.example.sendbacksendbag.ExpandableFabExample
+import com.example.sendbacksendbag.ui.messages.BlackHorizontalLine
+import com.example.sendbacksendbag.ui.messages.ExpandableFabExample
 
 @Composable
 fun HomeScreen(
